Add issue template
## Description

This PR introduces an issue template to the repository to help standardize and streamline the issue creation process. The template will guide contributors in providing the necessary information when submitting new issues.

## Type of change

Please delete options that are not relevant.

- [x] This change requires a documentation updat

## Additional Context (optional)
Adding an issue template ensures that all future issues will have a consistent format, making them easier to understand and manage. This will save time and effort for both the contributors and maintainers and improve the overall quality of the project.

## Checklist:

- [x] I have performed a self-review of my own code
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings (`make lint`)
- [ ] I have added tests that prove my fix is effective or that my feature works
- [ ] New and existing unit tests pass locally with my changes (`make test`)